refactor: extract Overview builders into focused dashboard_html_overview module
Move build_service_surface_html, build_priority_queue_html, and
build_overview_blocks (plus the _build_top_findings helper) from
dashboard_html_sections.py into a new dashboard_html_overview.py, matching
the Option B split pattern established on main.

dashboard_html_sections.py now acts as a thin re-export / delegator shim:
- 645 → 175 lines
- private names (_build_*) retained as 1-line delegators so dashboard-gen.py
  imports and all existing tests keep working with byte-identical HTML

Add 38 unit tests in test_dashboard_html_overview_unit.py covering each
builder's empty-fallback, populated happy path, HTML escaping, severity /
env_connected branch switches, and badge logic. Full suite: 553 passed
(was 515, +38).
